requirements.txt
addict==2.4.0
appdirs==1.4.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
contourpy==1.2.0
cycler==0.12.1
dill==0.3.6
dlib==19.24.1
docker-pycreds==0.4.0
face_alignment==1.3.4
filelock==3.13.1
fonttools==4.50.0
ftfy==6.2.0
gdown==3.12.2
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
imageio==2.34.0
joblib==1.2.0
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.42.0
matplotlib==3.7.1
networkx==3.2.1
numba==0.59.1
numpy==1.23.5
opencv-python==4.7.0.72
packaging==24.0
pandas==2.0.1
pathtools==0.1.2
Pillow==10.0.0
protobuf==4.25.3
psutil==5.9.8
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.43.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
threadpoolctl==3.4.0
tifffile==2024.2.12
torch==2.5.1
torchmetrics==0.11.4
torchvision==0.20.1
tqdm==4.66.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.15.10
wcwidth==0.2.13
ninja
grpcio